--- packages/react-select/src/animated/MultiValue.d.ts
@@ -1,21 +1,0 @@
-import { ComponentType } from 'react';
-import { MultiValueProps } from '../components/MultiValue';
-import { Collapse, fn } from './transitions';
-import { GroupTypeBase, OptionTypeBase } from '../types';
-
-export type AnimatedMultiValueProps<
-    OptionType extends OptionTypeBase,
-    GroupType extends GroupTypeBase<OptionType> = GroupTypeBase<OptionType>
-> = {
-    in?: boolean;
-    onExited?: fn;
-} & MultiValueProps<OptionType, GroupType>;
-
-export function AnimatedMultiValue<
-    OptionType extends OptionTypeBase,
-    GroupType extends GroupTypeBase<OptionType> = GroupTypeBase<OptionType>
->(
-    WrappedComponent: ComponentType<MultiValueProps<OptionType, GroupType>>,
-): ComponentType<AnimatedMultiValueProps<OptionType, GroupType>>;
-
-export default AnimatedMultiValue;
