+++ packages/react-select/src/animated/index.ts
