+++ packages/react-select/src/animated/transitions.tsx
